tests: fix aria practices tests not running
